pyproject.toml
[tool.poetry]
name = "botcommons"
version = "0.2.0"
description = ""
authors = ["JosXa"]

[tool.poetry.dependencies]
python = "^3.8"
pyrogram = {git = "https://github.com/pyrogram/pyrogram", rev = "asyncio-dev"}
tgintegration = {git = "https://github.com/josxa/tgintegration", rev = "async-develop"}
tgcrypto = "^1.2.0"
cached_property = "^1.5.1"
buslane = "^0.0.u"
python-decouple = "3.3"
logzero = "^.5.0"
more-itertools = "^8.2.0"
unsync = "^1.2.1"
haps = "^1.1.2"
typing-inspect = "^0.5.0"
pydantic = {extras = ["dotenv"], version = "^1.4"}
ordered_set = "^3.1.1"
decorators = "^0.1.1"
pyhumps = "^1.3.1"
redis-collections = "^0.8.1"
boltons = "^20.1.0"
pytz = "2020.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
pytest-asyncio = "^0.10.0"
black = "^19.10b0"
mypy = ">0.770"
birdseye = "^0.8.3"
mkinit = "^0.2.0"
typer = "^0.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
